feat: update page layout types in admin app
BREAKING CHANGE: Rename `PageWithConfig` to `PageWithLayout` to better reflect the purpose of the type

The `PageWithConfig` type has been renamed to `PageWithLayout` to better communicate that it's used for configuring the page layout, not the overall application configuration.

This change affects several files in the `admin` app, where the type is used to define the layout of individual pages.

BREAKING CHANGE: Remove `pageConfig.ts` module
Th...